+++ changelog.md
@@ -4,6 +4,10 @@ This changelog only includes the most important changes in recent updates. For a
 
 ## Version 3.x
 
+### Version 3.27.0
+* Reading SimulationArchives with version < 2 is no longer supported.
+* Internal changes on how SimulationArchives are written. 
+
 ### Version 3.26.3
 * A few more changes to reduce the number of compiler warnings. This should not affect any calculation.